+++ lang/fr/block_openveo_videos.php
@@ -1,14 +1,25 @@
 <?php
-// This file is part of OpenVeo - http://www.veo-labs.com/suite-logicielle-openveo
+// This file is part of Moodle - http://moodle.org/
 //
-// TODO : Add License information here
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle. If not, see <http://www.gnu.org/licenses/>.
 
 /**
  * Defines french translations of the plugin.
  *
  * @package block_openveo_videos
- * @copyright 2015, veo-labs <info@veo-labs.com>
- * @license TODO
+ * @copyright 2018 Veo-labs
+ * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 $string['pluginname'] = 'Vidéos OpenVeo';