+++ openveo_videos_settings.php
@@ -22,7 +22,7 @@
  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-require('../../../config.php');
+require('../../config.php');
 require_once($CFG->libdir . '/adminlib.php');
 
 use block_openveo_videos\output\settings_page;