+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - An OpenVeo icon is now displayed next to the plugin name in administration interface
 - Set player page title and header to the video title instead of the course name
 - Set manage page title (&lt;title&gt; tag)
+- Video thumbnails are now 200px width images instead of full size video images
 
 ## BUG FIXES