+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 5.0.1 / YYYY-MM-DD
+
+## BUG FIXES
+
+- Fix configuration page which didn't work at all
+
 # 5.0.0 / 2018-10-17
 
 ## BREAKING CHANGES